fixing to real

+++ wrapping/fortran/src/fortran/ext_fortran2.f90
@@ -1,6 +1,6 @@
 subroutine square(i, b)
-    integer, intent (in)  :: i              ! input
-    integer, intent (out) :: b              ! output
+    real, intent (in)  :: i              ! input
+    real, intent (out) :: b              ! output
 
     b = i**2
 end subroutine
